+++ src/DOTABATA_VRLand.Shared/Interfaces/StreamingHubs/IRoomHubReceiver.cs
@@ -156,7 +156,7 @@ namespace DOTABATA_VRLand.Shared.Interfaces.StreamingHubs {
         /// <summary>
         /// プレイヤーのステータス同期通知
         /// </summary>
-        public void OnSyncPlayerStatus(Guid playerConId, float hp);
+        public void OnSyncPlayerStatus(Guid playerConId, int hp);
     }
 
 }
